# Simplify decoding (no functional changes)