Hi! It looks like the original implementation back in https://github.com/trezor/trezor-crypto/commit/89a7d7797b806fac left a few too many sanity checks in the source. Adjusted and simplified to appease the go compiler.

Thank you for the excellent starting base, and the extensive test suite!

* Before
```
base58$ go test -bench=. -count=1 -benchmem ./
goos: darwin
goarch: amd64
pkg: github.com/mr-tron/base58
BenchmarkTrivialBase58Encoding-8   	  182612	      7054 ns/op	    1911 B/op	      90 allocs/op
BenchmarkFastBase58Encoding-8      	  928771	      1245 ns/op	     144 B/op	       2 allocs/op
BenchmarkTrivialBase58Decoding-8   	  359553	      3370 ns/op	     513 B/op	      48 allocs/op
BenchmarkFastBase58Decoding-8      	 1558393	       838 ns/op	     368 B/op	       3 allocs/op
PASS
ok  	github.com/mr-tron/base58	58.666s
```

* After ( when including https://github.com/mr-tron/base58/pull/15 )

```
base58$ go test -bench=. -count=1 -benchmem ./
goos: darwin
goarch: amd64
pkg: github.com/mr-tron/base58
BenchmarkTrivialBase58Encoding-8   	  155588	      7466 ns/op	    1911 B/op	      90 allocs/op
BenchmarkFastBase58Encoding-8      	  966726	      1245 ns/op	      96 B/op	       2 allocs/op
BenchmarkTrivialBase58Decoding-8   	  350648	      3447 ns/op	     513 B/op	      48 allocs/op
BenchmarkFastBase58Decoding-8      	 2276010	       523 ns/op	     127 B/op	       2 allocs/op
PASS
ok  	github.com/mr-tron/base58	57.326s
```

/cc @prusnak @Stebalien