@ribasushi 2020-05-24
> I don't have time today, but I will check and merge your pull requests tomorrow.

@mr-tron no worries, I clearly wasn't done anyway :) I **think** now I am done with tweaking all of this, the 3 PRs are good to look at, whenever time permits.

Thanks!