@ribasushi 2020-05-28
Hi there @mr-tron! Do you think you will have a chance to look at these before next Monday? Thanks!