@mr-tron 2020-05-29
I merged. You can also add your name into LICENSE file if you want.